qcacmn: Add 11BE wireless mode and NO11BE phybitmap
Extract the 11BE flags advertised by FW in WMI_SERVICE_READY_EXT2_EVENTID
and update the regulatory pdev wireless modes. The wireless mode is
currently represented as 'uint32_t' and all the 32 bits are exhausted.
To accommodate the new 11BE flags, change the data type of wireless_modes
from 'uint32_t' to 'uint64_t'.

Add a regulatory phymode 'REG_PHYMODE_11BE' in enum reg_phymode.

If the chip does not support 11BE phymode, FW advertises it via
WMI_REGULATORY_PHYBITMAP enum with a value of REGULATORY_PHYMODE_NO11BE.
Extract the phybitmap and store it in the regualtory
as 'REGULATORY_PHYMODE_NO11BE' bitmap.

CRs-Fixed: 2901260

+++ target_if/init_deinit/inc/service_ready_param.h
@@ -119,7 +119,7 @@ struct wlan_psoc_host_ppe_threshold {
  * @high_5ghz_chan: higher 5 GHz channels
  */
 struct wlan_psoc_host_hal_reg_cap_ext {
-	uint32_t wireless_modes;
+	uint64_t wireless_modes;
 	uint32_t low_2ghz_chan;
 	uint32_t high_2ghz_chan;
 	uint32_t low_5ghz_chan;
@@ -263,7 +263,7 @@ struct wlan_psoc_host_mac_phy_caps_ext2 {
 	uint32_t hw_mode_id;
 	uint32_t pdev_id;
 	uint32_t phy_id;
-	uint32_t wireless_modes_ext;
+	uint64_t wireless_modes_ext;
 #ifdef WLAN_FEATURE_11BE
 	uint32_t eht_cap_info_2G[PSOC_HOST_MAX_MAC_SIZE];
 	uint32_t eht_supp_mcs_2G;
