+++ umac/regulatory/dispatcher/inc/wlan_reg_ucfg_api.h
@@ -400,7 +400,7 @@ QDF_STATUS ucfg_reg_set_hal_reg_cap(struct wlan_objmgr_psoc *psoc,
  * Return: QDF_STATUS
  */
 QDF_STATUS ucfg_reg_update_hal_reg_cap(struct wlan_objmgr_psoc *psoc,
-				       uint32_t wireless_modes, uint8_t phy_id);
+				       uint64_t wireless_modes, uint8_t phy_id);
 
 /**
  * ucfg_set_ignore_fw_reg_offload_ind() - API to set ignore regdb offload ind
